Overhaul file structure & add NEO motors
