+++ swervelib/abstract/__init__.py
@@ -1,0 +1,19 @@
+__all__ = [
+    "SendableABCMeta",
+    "CoaxialDriveComponent",
+    "CoaxialAzimuthComponent",
+    "Gyro",
+    "AbsoluteEncoder",
+    "SwerveModule",
+]
+
+from abc import ABCMeta
+from wpiutil import Sendable
+
+from motor import CoaxialDriveComponent, CoaxialAzimuthComponent
+from sensor import Gyro, AbsoluteEncoder
+from system import SwerveModule
+
+
+class SendableABCMeta(ABCMeta, type(Sendable)):
+    pass
